refactor(api): add validator middelware and remove validations from route and controller
