+++ controllers/authControllers.js
@@ -1,16 +1,8 @@
-const {validationResult} = require('express-validator');
 const {User} = require('../models');
 const bcrypt = require('bcrypt');
 
 const register = (req, res, next) => {
   try{
-    // Get the validation result
-    const errors = validationResult(req);
-    if(!errors.isEmpty()){
-      return res.status(400).json({
-        errors: errors.array()
-      });
-    }
     // Get the user data
     const {firstName, lastName, email, password} = req.body;
     // Hash the password
